Move the skeleton to payment-processor.js
